refactor: remove unused models and middleware, add authentication middleware
