+++ backend/middlewares/verifyToken.js
@@ -1,8 +1,5 @@
 const jwt = require("jsonwebtoken");
 
-/**
- * Middleware to verify JWT token and authenticate user
- */
 const verifyToken = (req, res, next) => {
   try {
     const authHeader = req.headers.authorization;
