--- backend/models/databasePreconfigs.js
@@ -1,13 +1,0 @@
-
-const { superAdminController } = require('../routes/superAdminRouters');
-
-const startupFunctions = () => {
-
-    // Create a super admin
-    // superAdminController.createUniversity(
-    //     'university@gmail.com',
-    //     'uni123'
-    // );
-}
-
-module.exports = startupFunctions;
