--- backend/models/universityModel.js
@@ -1,9 +1,0 @@
-const mongoose = require('mongoose');
-
-const universitySchema = new mongoose.Schema({
-  name: { type: String, required: true },
-  email: { type: String,  required: true },
-  password: { type: String, required: true },
-}, { timestamps: true });
-
-module.exports = mongoose.model('university', universitySchema);
