+++ backend/utils/notification.js
@@ -1,7 +1,5 @@
 const sendMessageToUser = require("..");
-const departmentModel = require("../models/departmentModel");
 const notificationModel = require("../models/notificationModel");
-const studentModel = require("../models/studentModel");
 const { sendEmail } = require("./email")
 
 const notification = {
@@ -19,63 +17,63 @@ const notification = {
         }
     },
     notifiyUserById: async (userId, message) => {
-        const user = await studentModel.findById(userId);
-        if (user) {
+        // const user = await studentModel.findById(userId);
+        // if (user) {
             sendMessageToUser(userId, message);
             this.appendNotification(userId, message, 'student');
-            if (user.allowEmails) {
-                sendEmail(user.email, message);
-            }
-        }
+            // if (user.allowEmails) {
+            //     sendEmail(user.email, message);
+            // }
+        // }
     },
 
-    notifyAllUsers: async (message) => {
-        try {
-            const students = await studentModel.find({ allowEmails: true });
-            students.forEach(student => {
-                sendMessageToUser(student._id, message);
-                this.appendNotification(student._id, message, 'student');
-                if (student.allowEmails) {
-                    sendEmail(student.email, message);
-                }
-            });
-        } catch (error) {
-            console.log("Error at notifyAllUsers", error);
-        }
-    },
+    // notifyAllUsers: async (message) => {
+    //     try {
+    //         const students = await studentModel.find({ allowEmails: true });
+    //         students.forEach(student => {
+    //             sendMessageToUser(student._id, message);
+    //             this.appendNotification(student._id, message, 'student');
+    //             // if (student.allowEmails) {
+    //             //     sendEmail(student.email, message);
+    //             // }
+    //         });
+    //     } catch (error) {
+    //         console.log("Error at notifyAllUsers", error);
+    //     }
+    // },
 
-    notifyUserByBranch: async (branch, message) => {
-        try {
-            const students = await studentModel.find({ branch, allowEmails: true });
-            students.forEach(student => {
-                sendMessageToUser(student._id, message);
-                this.appendNotification(student._id, message, 'student');
-                if (student.allowEmails) {
-                    sendEmail(student.email, message);
-                }
-            });
-        } catch (error) {
-            console.log("Error at notifyUserByBranch", error);
-        }
-    },
+    // notifyUserByBranch: async (branch, message) => {
+    //     try {
+    //         const students = await studentModel.find({ branch, allowEmails: true });
+    //         students.forEach(student => {
+    //             sendMessageToUser(student._id, message);
+    //             this.appendNotification(student._id, message, 'student');
+    //             // if (student.allowEmails) {
+    //             //     sendEmail(student.email, message);
+    //             // }
+    //         });
+    //     } catch (error) {
+    //         console.log("Error at notifyUserByBranch", error);
+    //     }
+    // },
 
-    notifyFaculty: async (facultyIds, message) => {
-        try {
-            const departments = await departmentModel.find({
-                _id: { $in: facultyIds },
-                allowEmails: true
-            });
-            departments.forEach(department => {
-                sendMessageToUser(department.id, message);
-                this.appendNotification(department.id, message, 'faculty');
-                if (department.allowEmails) {
-                    sendEmail(department.email, message);
-                }
-            });
-        } catch (error) {
-            console.log("Error at notifyFaculty", error);
-        }
-    }
+    // notifyFaculty: async (facultyIds, message) => {
+    //     try {
+    //         const departments = await departmentModel.find({
+    //             _id: { $in: facultyIds },
+    //             allowEmails: true
+    //         });
+    //         departments.forEach(department => {
+    //             sendMessageToUser(department.id, message);
+    //             this.appendNotification(department.id, message, 'faculty');
+    //             // if (department.allowEmails) {
+    //             //     sendEmail(department.email, message);
+    //             // }
+    //         });
+    //     } catch (error) {
+    //         console.log("Error at notifyFaculty", error);
+    //     }
+    // }
 };
 
 module.exports = notification;
